Move some classes to different files
